Move default manifest config.
