+++ crates/core/Cargo.toml
@@ -18,7 +18,6 @@ version_spec = { version = "0.2.0", path = "../version-spec", features = [
 ] }
 warpgate = { version = "0.9.2", path = "../warpgate", features = ["schematic"] }
 cached = { workspace = true }
-extism = { workspace = true }
 human-sort = { workspace = true }
 indexmap = { workspace = true }
 miette = { workspace = true }
