+++ crates/core/src/helpers.rs
@@ -19,6 +19,7 @@ use std::{io, thread};
 use tracing::trace;
 
 pub static ENV_VAR: Lazy<Regex> = Lazy::new(|| Regex::new(r"\$(?<name>[A-Z0-9_]+)").unwrap());
+
 pub static ENV_VAR_SUB: Lazy<Regex> =
     Lazy::new(|| Regex::new(r"\$\{(?<name>[A-Z0-9_]+)\}").unwrap());
 
@@ -28,10 +29,6 @@ pub fn get_proto_version() -> Version {
 }
 
 pub fn get_proto_home() -> miette::Result<PathBuf> {
-    // if cfg!(debug_assertions) {
-    //     return Ok(get_home_dir()?.join(".proto-debug"));
-    // }
-
     if let Ok(root) = env::var("PROTO_HOME") {
         return Ok(root.into());
     }
