+++ crates/pdk-test-utils/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/moonrepo/proto"
 [dependencies]
 proto_core = { version = "0.28.4", path = "../core" }
 proto_pdk_api = { version = "0.14.2", path = "../pdk-api" }
+warpgate = { version = "0.9.2", path = "../warpgate" }
 extism = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
