+++ crates/pdk-test-utils/src/lib.rs
@@ -4,18 +4,20 @@ mod wrapper;
 pub use proto_core as core;
 pub use proto_core::{
     Id, ProtoEnvironment, Tool, ToolManifest, UnresolvedVersionSpec, Version, VersionReq,
-    VersionSpec, Wasm,
+    VersionSpec,
 };
 pub use proto_pdk_api::*;
+pub use warpgate::Wasm;
 pub use wrapper::WasmTestWrapper;
 
-use proto_core::{inject_default_manifest_config, inject_proto_manifest_config};
+use proto_core::inject_proto_manifest_config;
 use serde::Serialize;
 use std::collections::HashMap;
 use std::fs::OpenOptions;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::{env, fs};
+use warpgate::inject_default_manifest_config;
 
 static mut LOGGING: bool = false;
