+++ crates/warpgate/src/lib.rs
@@ -16,5 +16,6 @@ pub use loader::*;
 pub use locator::*;
 pub use plugin::*;
 
+pub use extism::{Manifest as PluginManifest, Wasm};
 pub use warpgate_api as api;
 pub use warpgate_api::VirtualPath;
